# Rewrite using channels and a worker pool