--- a/go.mod
+++ b/go.mod
@@ -6,6 +6,7 @@ require (
 	github.com/mattn/go-colorable v0.1.12
 	go.uber.org/zap v1.21.0
 	golang.org/x/exp v0.0.0-20220713135740-79cabaa25d75
+	golang.org/x/sync v0.0.0-20210220032951-036812b2e83c
 )
 
 require (