--- a/pkg/downloader/downloader_test.go
+++ b/pkg/downloader/downloader_test.go
@@ -2,11 +2,13 @@ package downloader
 
 import (
 	"os"
+	"path"
+	"strconv"
 	"testing"
 )
 
 func TestDownload(t *testing.T) {
-	cfg := DownloaderSettings{
+	s := Settings{
 		Verbose:      false,
 		ShowProgress: false,
 		IncludeVideo: false,
@@ -19,12 +21,30 @@ func TestDownload(t *testing.T) {
 		MinHeight:    0,
 	}
 
-	count, err := Download(cfg, Filters)
+	count, err := Download(s, Filters)
 	if err != nil {
-		t.Fatalf("Download(%#v) error: %v", cfg, err)
+		t.Fatalf("Download(%#v) error: %v", s, err)
 	}
 
-	if count != uint32(cfg.Count) {
-		t.Fatalf("Download(%#v) loaded %v media, expected %v", cfg, count, cfg.Count)
+	if count != int64(s.Count) {
+		t.Fatalf("Download(%#v) loaded %v media, expected %v", s, count, s.Count)
+	}
+}
+
+func BenchmarkDownload(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		s := Settings{
+			Verbose:      false,
+			ShowProgress: false,
+			IncludeVideo: true,
+			Subreddit:    "wallpaper",
+			Sorting:      "best",
+			Timeframe:    "all",
+			Directory:    path.Join(os.TempDir(), strconv.Itoa(i)),
+			Count:        35,
+			MinWidth:     1920,
+			MinHeight:    1080,
+		}
+		Download(s, Filters)
 	}
 }