--- a/pkg/downloader/filters.go
+++ b/pkg/downloader/filters.go
@@ -7,32 +7,32 @@ var Filters = []Filter{whFilter, urlFilter}
 
 // Interface that filters the given slice and returns the mutated version of it.
 type Filter interface {
-	Filter([]toDownload, *DownloaderSettings) []toDownload
+	Filter([]content, *Settings) []content
 }
 
 // []downloadable according to its own logic.
 // FilterFunc implements filter interface and expects the function to return a new slice.
-type FilterFunc func([]toDownload, *DownloaderSettings) []toDownload
+type FilterFunc func([]content, *Settings) []content
 
-func (f FilterFunc) Filter(td []toDownload, ds *DownloaderSettings) []toDownload {
-	return f(td, ds)
+func (f FilterFunc) Filter(c []content, s *Settings) []content {
+	return f(c, s)
 }
 
 var (
-	whFilter FilterFunc = func(td []toDownload, ds *DownloaderSettings) []toDownload {
-		f := make([]toDownload, 0)
-		for _, m := range td {
-			if m.Data.Width >= ds.MinWidth && m.Data.Height >= ds.MinHeight {
+	whFilter FilterFunc = func(c []content, s *Settings) []content {
+		f := make([]content, 0)
+		for _, m := range c {
+			if m.Width >= s.MinWidth && m.Height >= s.MinHeight {
 				f = append(f, m)
 			}
 		}
 		return f
 	}
 
-	urlFilter FilterFunc = func(td []toDownload, ds *DownloaderSettings) []toDownload {
-		f := make([]toDownload, 0)
-		for _, m := range td {
-			if len(m.Data.URL) > 0 && utils.IsURL(m.Data.URL) {
+	urlFilter FilterFunc = func(c []content, s *Settings) []content {
+		f := make([]content, 0)
+		for _, m := range c {
+			if len(m.URL) > 0 && utils.IsURL(m.URL) {
 				f = append(f, m)
 			}
 		}