Now, we have the following structure:

1. Single goroutine fetches media and sends it to a channel.

2. 16 goroutines watch the queue and download media from it,
after downloading, they send the result to the file queue.

3. Single goroutine waits for items in file queue and
saves each file to disk sequentially.

This makes downloads happen faster if there are multiple images in my tests.

Before:
goos: windows
goarch: amd64
pkg: redditdl/pkg/downloader
cpu: AMD Ryzen 5 3600 6-Core Processor              
BenchmarkDownload-12             1  19739944800 ns/op   9959408 B/op     51208 allocs/op
PASS
ok    redditdl/pkg/downloader  19.862s

After:
goos: windows
goarch: amd64
pkg: redditdl/pkg/downloader
cpu: AMD Ryzen 5 3600 6-Core Processor              
BenchmarkDownload-12             1  12474094500 ns/op  296305624 B/op     31833 allocs/op
PASS
ok    redditdl/pkg/downloader  12.688s